pyproject.toml
[project]
name = "buying_winners"
version = "0.1.0"
description = "buying_winners"
authors = [
    {name = "Le Quang Hai",email = "lehai0609@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.11.9"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

